Move RegistoryFS to RegistryFS Fixes #40
