+++ DokanNet.Tests/Properties/AssemblyInfo.cs
