+++ sample/RegistryFS/app.config
